dapp-refactor: moved handleRetrievingTransactionResultsFromLocalStorage and handleSanitizeFormInputs to main common shared method folder
