+++ system-contract-dapp-playground/src/components/common/methods/handleRetrievingTransactionResultsFromLocalStorage.tsx
@@ -19,11 +19,11 @@
  */
 
 import { Dispatch, SetStateAction } from 'react';
-import { TRANSACTION_PAGE_SIZE } from '../states/commonStates';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { TransactionResult } from '@/types/contract-interactions/HTS';
 import { getArrayTypedValuesFromLocalStorage } from '@/api/localStorage';
 import { HEDERA_COMMON_WALLET_REVERT_REASONS } from '@/utils/common/constants';
+import { TRANSACTION_PAGE_SIZE } from '../../contract-interaction/hts/shared/states/commonStates';
 
 export const handleRetrievingTransactionResultsFromLocalStorage = (
   toaster: any,
