+++ system-contract-dapp-playground/src/components/common/methods/handleSanitizeFormInputs.tsx
