+++ system-contract-dapp-playground/src/components/contract-interaction/exchange-rate/methods/index.tsx
@@ -24,17 +24,17 @@ import { useToast } from '@chakra-ui/react';
 import { useEffect, useState } from 'react';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { TransactionResult } from '@/types/contract-interactions/HTS';
-import { handleAPIErrors } from '../../hts/shared/methods/handleAPIErrors';
+import { handleAPIErrors } from '../../../common/methods/handleAPIErrors';
+import { useToastSuccessful } from '../../../../hooks/useToastSuccessful';
 import { handleExchangeRate } from '@/api/hedera/exchange-rate-interactions';
 import { TRANSACTION_PAGE_SIZE } from '../../hts/shared/states/commonStates';
-import { useToastSuccessful } from '../../../../hooks/useToastSuccessful';
 import { usePaginatedTxResults } from '../../../../hooks/usePaginatedTxResults';
 import { TransactionResultTable } from '../../hts/shared/components/TransactionResultTable';
-import { handleSanitizeHederaFormInputs } from '../../hts/shared/methods/handleSanitizeFormInputs';
-import { SharedExecuteButton, SharedFormInputField } from '../../hts/shared/components/ParamInputForm';
+import { handleSanitizeHederaFormInputs } from '../../../common/methods/handleSanitizeFormInputs';
 import { useUpdateTransactionResultsToLocalStorage } from '../../../../hooks/useUpdateLocalStorage';
+import { SharedExecuteButton, SharedFormInputField } from '../../hts/shared/components/ParamInputForm';
 import useFilterTransactionsByContractAddress from '../../../../hooks/useFilterTransactionsByContractAddress';
-import { handleRetrievingTransactionResultsFromLocalStorage } from '../../hts/shared/methods/handleRetrievingTransactionResultsFromLocalStorage';
+import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
 import {
   CONTRACT_NAMES,
   HEDERA_BRANDING_COLORS,
