+++ system-contract-dapp-playground/src/components/contract-interaction/ihrc/methods/index.tsx
@@ -25,14 +25,14 @@ import { getWalletProvider } from '@/api/wallet';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { handleIHRC719APIs } from '@/api/hedera/ihrc-interactions';
 import { TransactionResult } from '@/types/contract-interactions/HTS';
-import { handleAPIErrors } from '../../hts/shared/methods/handleAPIErrors';
-import { TRANSACTION_PAGE_SIZE } from '../../hts/shared/states/commonStates';
+import { handleAPIErrors } from '../../../common/methods/handleAPIErrors';
 import { useToastSuccessful } from '../../../../hooks/useToastSuccessful';
+import { TRANSACTION_PAGE_SIZE } from '../../hts/shared/states/commonStates';
 import { usePaginatedTxResults } from '../../../../hooks/usePaginatedTxResults';
 import { TransactionResultTable } from '../../hts/shared/components/TransactionResultTable';
-import { SharedExecuteButton, SharedFormInputField } from '../../hts/shared/components/ParamInputForm';
 import { useUpdateTransactionResultsToLocalStorage } from '../../../../hooks/useUpdateLocalStorage';
-import { handleRetrievingTransactionResultsFromLocalStorage } from '../../hts/shared/methods/handleRetrievingTransactionResultsFromLocalStorage';
+import { SharedExecuteButton, SharedFormInputField } from '../../hts/shared/components/ParamInputForm';
+import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
 import {
   HEDERA_TRANSACTION_RESULT_STORAGE_KEYS,
   HEDERA_BRANDING_COLORS,
